+++ packages/riverpod/test/new/providers/future_provider/auto_dispose_future_provider_test.dart
@@ -6,7 +6,7 @@ import 'package:riverpod/riverpod.dart';
 import 'package:riverpod/src/internals.dart' show ProviderElementBase;
 import 'package:test/test.dart';
 
-import '../../utils.dart';
+import '../../../old/utils.dart';
 
 void main() {
   test('can read and set current AsyncValue', () {
