+++ packages/riverpod/test/new/providers/provider/auto_dispose_provider_family_test.dart
@@ -3,7 +3,7 @@ import 'package:riverpod/riverpod.dart';
 import 'package:riverpod/src/internals.dart' show ProviderElementBase;
 import 'package:test/test.dart';
 
-import '../../utils.dart';
+import '../../../old/utils.dart';
 
 void main() {
   group('Provider.autoDispose.family', () {
