+++ packages/riverpod/test/new/providers/stream_notifier/stream_notifier_test.dart
@@ -1,12 +1,15 @@
+// ignore_for_file: avoid_types_on_closure_parameters
+
 import 'dart:async';
 
 import 'package:meta/meta.dart';
 import 'package:mockito/mockito.dart';
-import 'package:riverpod/src/internals.dart';
+import 'package:riverpod/legacy.dart';
+import 'package:riverpod/riverpod.dart';
 import 'package:test/test.dart';
 
-import '../../../new/matrix.dart';
-import '../../utils.dart';
+import '../../matrix.dart';
+import '../../../old/utils.dart';
 
 void main() {
   test('Throws if using notifier properties in its constructor', () {
@@ -20,10 +23,10 @@ void main() {
     );
   });
 
-  asyncNotifierProviderFactory.createGroup((factory) {
+  streamNotifierProviderFactory.createGroup((factory) {
     test('Cannot share a Notifier instance between providers ', () {
       final container = ProviderContainer.test();
-      final notifier = factory.deferredNotifier<int>((ref) => 0);
+      final notifier = factory.deferredNotifier((ref) => Stream.value(0));
 
       final provider = factory.provider<int>(() => notifier);
       final provider2 = factory.provider<int>(() => notifier);
@@ -38,12 +41,12 @@ void main() {
 
     test('Can read state inside onDispose', () {
       final container = ProviderContainer.test();
-      late TestAsyncNotifier<int> notifier;
+      late TestStreamNotifier<int> notifier;
       final provider = factory.simpleTestProvider((ref) {
         ref.onDispose(() {
           notifier.state;
         });
-        return 0;
+        return Stream.value(0);
       });
 
       container.listen(provider.notifier, (prev, next) {});
@@ -54,7 +57,7 @@ void main() {
 
     test('Using the notifier after dispose throws', () {
       final container = ProviderContainer.test();
-      final provider = factory.simpleTestProvider((ref) => 0);
+      final provider = factory.simpleTestProvider((ref) => Stream.value(0));
 
       container.listen(provider.notifier, (prev, next) {});
       final notifier = container.read(provider.notifier);
@@ -69,68 +72,6 @@ void main() {
       expect(() => notifier.update((p1) => 42), throwsStateError);
     });
 
-    test('Can assign `AsyncLoading<T>` to `AsyncValue<void>`', () {
-      // Regression test for https://github.com/rrousselGit/riverpod/issues/2120
-      final provider = factory.simpleTestProvider<void>((ref) => 42);
-      final container = ProviderContainer.test();
-
-      final sub = container.listen(provider.notifier, (prev, next) {});
-
-      // ignore: void_checks
-      expect(sub.read().state, const AsyncData<void>(42));
-
-      sub.read().state = const AsyncLoading<int>();
-
-      expect(
-        sub.read().state,
-        isA<AsyncLoading<void>>()
-            .having((e) => e.hasValue, 'hasValue', true)
-            .having((e) => e.value, 'value', 42),
-      );
-    });
-
-    test('Can assign `AsyncData<T>` to `AsyncValue<void>`', () {
-      // Regression test for https://github.com/rrousselGit/riverpod/issues/2120
-      final provider = factory.simpleTestProvider<void>((ref) => 42);
-      final container = ProviderContainer.test();
-
-      final sub = container.listen(provider.notifier, (prev, next) {});
-
-      // ignore: void_checks
-      expect(sub.read().state, const AsyncData<void>(42));
-
-      sub.read().state = const AsyncData<int>(42);
-
-      expect(
-        sub.read().state,
-        isA<AsyncData<void>>()
-            .having((e) => e.hasValue, 'hasValue', true)
-            .having((e) => e.value, 'value', 42),
-      );
-    });
-
-    test('Can assign `AsyncError<T>` to `AsyncValue<void>`', () {
-      // Regression test for https://github.com/rrousselGit/riverpod/issues/2120
-      final provider = factory.simpleTestProvider<void>((ref) => 42);
-      final container = ProviderContainer.test();
-
-      final sub = container.listen(provider.notifier, (prev, next) {});
-
-      // ignore: void_checks
-      expect(sub.read().state, const AsyncData<void>(42));
-
-      sub.read().state = AsyncError<int>(21, StackTrace.current);
-
-      expect(
-        sub.read().state,
-        isA<AsyncError<void>>()
-            .having((e) => e.hasValue, 'hasValue', true)
-            .having((e) => e.value, 'value', 42)
-            .having((e) => e.hasError, 'hasError', true)
-            .having((e) => e.error, 'error', 21),
-      );
-    });
-
     group('supports AsyncValue transition', () {
       test(
           'performs seamless copyWithPrevious if triggered by ref.invalidate/ref.refresh',
@@ -138,7 +79,7 @@ void main() {
         final container = ProviderContainer.test();
         var count = 0;
         final provider = factory.simpleTestProvider(
-          (ref) => Future.value(count++),
+          (ref) => Stream.value(count++),
         );
 
         container.listen(provider, (previous, next) {});
@@ -168,7 +109,7 @@ void main() {
           'performs seamless:false copyWithPrevious on `state = AsyncLoading()`',
           () async {
         final container = ProviderContainer.test();
-        final provider = factory.simpleTestProvider((ref) => Future.value(0));
+        final provider = factory.simpleTestProvider((ref) => Stream.value(0));
 
         final sub = container.listen(provider.notifier, (previous, next) {});
 
@@ -190,7 +131,7 @@ void main() {
         final container = ProviderContainer.test();
         final dep = StateProvider((ref) => 0);
         final provider = factory.simpleTestProvider(
-          (ref) => Future.value(ref.watch(dep)),
+          (ref) => Stream.value(ref.watch(dep)),
         );
 
         container.listen(provider, (previous, next) {});
@@ -209,39 +150,13 @@ void main() {
         expect(container.read(provider), const AsyncData(1));
       });
 
-      test('performs seamless data > loading > error transition', () async {
-        final container = ProviderContainer.test();
-        var result = Future.value(42);
-        final provider = FutureProvider((ref) => result);
-
-        final sub = container.listen(provider.future, (_, __) {});
-
-        expect(container.read(provider), const AsyncLoading<int>());
-        expect(await sub.read(), 42);
-        expect(container.read(provider), const AsyncData<int>(42));
-
-        result = Future.error('err', StackTrace.empty);
-        container.invalidate(provider);
-
-        expect(
-          container.read(provider),
-          const AsyncLoading<int>().copyWithPrevious(const AsyncData(42)),
-        );
-        await expectLater(sub.read(), throwsA('err'));
-        expect(
-          container.read(provider),
-          const AsyncError<int>('err', StackTrace.empty)
-              .copyWithPrevious(const AsyncData<int>(42)),
-        );
-      });
-
       test(
           'performs seamless:false copyWithPrevious if both triggered by a dependency change and ref.refresh',
           () async {
         final container = ProviderContainer.test();
         final dep = StateProvider((ref) => 0);
         final provider = factory.simpleTestProvider(
-          (ref) => Future.value(ref.watch(dep)),
+          (ref) => Stream.value(ref.watch(dep)),
         );
 
         container.listen(provider, (previous, next) {});
@@ -262,7 +177,7 @@ void main() {
     });
 
     test('does not notify listeners when refreshed during loading', () async {
-      final provider = factory.simpleTestProvider((ref) => Future.value(0));
+      final provider = factory.simpleTestProvider((ref) => Stream.value(0));
       final container = ProviderContainer.test();
       final listener = Listener<AsyncValue<int>>();
 
@@ -310,9 +225,9 @@ void main() {
     });
 
     test(
-        'converts AsyncNotifier.build into an AsyncData if the future completes',
+        'converts StreamNotifier.build into an AsyncData if the future completes',
         () async {
-      final provider = factory.simpleTestProvider((ref) => Future.value(0));
+      final provider = factory.simpleTestProvider((ref) => Stream.value(0));
       final container = ProviderContainer.test();
       final listener = Listener<AsyncValue<int>>();
 
@@ -336,10 +251,10 @@ void main() {
       );
     });
 
-    test('converts AsyncNotifier.build into an AsyncError if the future fails',
+    test('converts StreamNotifier.build into an AsyncError if the future fails',
         () async {
       final provider = factory.simpleTestProvider<int>(
-        (ref) => Future.error(0, StackTrace.empty),
+        (ref) => Stream.error(0, StackTrace.empty),
       );
       final container = ProviderContainer.test();
       final listener = Listener<AsyncValue<int>>();
@@ -364,7 +279,8 @@ void main() {
       );
     });
 
-    test('supports cases where the AsyncNotifier constructor throws', () async {
+    test('supports cases where the StreamNotifier constructor throws',
+        () async {
       final provider = factory.provider<int>(
         () => Error.throwWithStackTrace(0, StackTrace.empty),
       );
@@ -386,21 +302,7 @@ void main() {
     });
 
     test(
-        'synchronously emits AsyncData if AsyncNotifier.build emits synchronously',
-        () async {
-      final provider = factory.simpleTestProvider<int>((ref) => 0);
-      final container = ProviderContainer.test();
-      final listener = Listener<AsyncValue<int>>();
-
-      container.listen(provider, listener.call, fireImmediately: true);
-
-      verifyOnly(listener, listener(null, const AsyncData(0)));
-      expect(container.read(provider.notifier).state, const AsyncData(0));
-      await expectLater(container.read(provider.future), completion(0));
-    });
-
-    test(
-        'synchronously emits AsyncError if AsyncNotifier.build throws synchronously',
+        'synchronously emits AsyncError if StreamNotifier.build throws synchronously',
         () async {
       final provider = factory.simpleTestProvider<int>(
         (ref) => Error.throwWithStackTrace(42, StackTrace.empty),
@@ -431,7 +333,7 @@ void main() {
         1: Completer<int>.sync(),
       };
       final provider = factory.simpleTestProvider<int>(
-        (ref) => completers[ref.watch(dep)]!.future,
+        (ref) => Stream.fromFuture(completers[ref.watch(dep)]!.future),
       );
       final listener = Listener<AsyncValue<int>>();
 
@@ -470,7 +372,7 @@ void main() {
         1: Completer<int>.sync(),
       };
       final provider = factory.simpleTestProvider<int>(
-        (ref) => completers[ref.watch(dep)]!.future,
+        (ref) => Stream.fromFuture(completers[ref.watch(dep)]!.future),
       );
       final listener = Listener<AsyncValue<int>>();
 
@@ -502,14 +404,17 @@ void main() {
       );
     });
 
-    group('AsyncNotifier.state', () {
+    group('StreamNotifier.state', () {
       test(
           'when manually modifying the state, the new exposed value contains the previous state when possible',
           () async {
-        final provider = factory.simpleTestProvider<int>((ref) => 0);
+        final provider = factory.simpleTestProvider<int>(
+          (ref) => Stream.value(0),
+        );
         final container = ProviderContainer.test();
 
         final sub = container.listen(provider.notifier, (previous, next) {});
+        await container.read(provider.future);
 
         // ignore: prefer_const_constructors, not using `const` as we voluntarily break identity to test `identical`
         final newState = AsyncData(84);
@@ -549,7 +454,7 @@ void main() {
         final provider = factory.simpleTestProvider<int>(
           (ref) {
             listener();
-            return Future.value(ref.watch(dep));
+            return Stream.value(ref.watch(dep));
           },
         );
         final container = ProviderContainer.test();
@@ -579,11 +484,11 @@ void main() {
         late AsyncValue<int> state;
         final provider = factory.provider<int>(
           () {
-            late TestAsyncNotifier<int> notifier;
+            late TestStreamNotifier<int> notifier;
             return notifier = factory.deferredNotifier<int>(
               (ref) {
                 state = notifier.state;
-                return Future.value(ref.watch(dep));
+                return Stream.value(ref.watch(dep));
               },
             );
           },
@@ -603,10 +508,13 @@ void main() {
         );
       });
 
-      test('notifies listeners when the setter is called', () {
-        final provider = factory.simpleTestProvider((ref) => 0);
+      test('notifies listeners when the setter is called', () async {
+        final provider = factory.simpleTestProvider((ref) => Stream.value(0));
         final container = ProviderContainer.test();
         final listener = Listener<AsyncValue<int>>();
+
+        // Skip the loading
+        await container.listen(provider.future, (previous, next) {}).read();
 
         container.listen(provider, listener.call);
 
@@ -621,14 +529,14 @@ void main() {
       });
     });
 
-    group('AsyncNotifier.future', () {
+    group('StreamNotifier.future', () {
       test(
-          'when disposed during loading, resolves with the content of AsyncNotifier.build',
+          'when disposed during loading, resolves with the content of StreamNotifier.build',
           () async {
         final container = ProviderContainer.test();
         final completer = Completer<int>.sync();
         final provider = factory.simpleTestProvider<int>(
-          (ref) => completer.future,
+          (ref) => Stream.fromFuture(completer.future),
         );
 
         final future = container.read(provider.future);
@@ -640,12 +548,12 @@ void main() {
       });
 
       test(
-          'when disposed during loading, resolves with the error of AsyncNotifier.build',
+          'when disposed during loading, resolves with the error of StreamNotifier.build',
           () async {
         final container = ProviderContainer.test();
         final completer = Completer<int>.sync();
         final provider = factory.simpleTestProvider<int>(
-          (ref) => completer.future,
+          (ref) => Stream.fromFuture(completer.future),
         );
 
         final future = container.read(provider.future);
@@ -664,7 +572,7 @@ void main() {
           final container = ProviderContainer.test();
           final completer = Completer<int>.sync();
           final provider = factory.simpleTestProvider<int>(
-            (ref) => completer.future,
+            (ref) => Stream.fromFuture(completer.future),
           );
 
           container.read(provider);
@@ -687,7 +595,7 @@ void main() {
           final container = ProviderContainer.test();
           final completer = Completer<int>.sync();
           final provider = factory.simpleTestProvider<int>(
-            (ref) => completer.future,
+            (ref) => Stream.fromFuture(completer.future),
           );
 
           container.read(provider);
@@ -706,12 +614,12 @@ void main() {
       );
 
       test(
-          'resolves with the new state if AsyncNotifier.state is modified during loading',
+          'resolves with the new state if StreamNotifier.state is modified during loading',
           () async {
         final container = ProviderContainer.test();
         final completer = Completer<int>.sync();
         final provider = factory.simpleTestProvider<int>(
-          (ref) => completer.future,
+          (ref) => Stream.fromFuture(completer.future),
         );
         final listener = Listener<Future<int>>();
 
@@ -736,7 +644,9 @@ void main() {
       test('resolves with the new state when notifier.state is changed',
           () async {
         final container = ProviderContainer.test();
-        final provider = factory.simpleTestProvider<int>((ref) => 0);
+        final provider = factory.simpleTestProvider<int>(
+          (ref) => Stream.value(0),
+        );
         final listener = Listener<Future<int>>();
 
         final sub = container.listen(provider.notifier, (previous, next) {});
@@ -763,7 +673,7 @@ void main() {
         final provider = factory.simpleTestProvider<int>(
           (ref) {
             listener();
-            return Future.value(ref.watch(dep));
+            return Stream.value(ref.watch(dep));
           },
         );
         final container = ProviderContainer.test();
@@ -782,7 +692,7 @@ void main() {
         final provider = factory.simpleTestProvider<int>(
           (ref) {
             listener();
-            return Future.value(ref.watch(dep));
+            return Stream.value(ref.watch(dep));
           },
         );
         final container = ProviderContainer.test();
@@ -802,7 +712,7 @@ void main() {
       });
     });
 
-    group('AsyncNotifierProvider.notifier', () {
+    group('StreamNotifierProvider.notifier', () {
       test(
           'never emits an update. The Notifier is never recreated once it is instantiated',
           () async {
@@ -810,9 +720,14 @@ void main() {
         final dep = StateProvider((ref) => 0);
         final provider = factory.provider<int>(() {
           listener();
-          return factory.deferredNotifier<int>((ref) => ref.watch(dep));
+          return factory.deferredNotifier(
+            (ref) => Stream.value(ref.watch(dep)),
+          );
         });
         final container = ProviderContainer.test();
+
+        // Skip the loading
+        await container.listen(provider.future, (previous, next) {}).read();
 
         container.listen(provider, (previous, next) {});
         final notifier = container.read(provider.notifier);
@@ -821,6 +736,7 @@ void main() {
         expect(container.read(provider), const AsyncData(0));
 
         container.read(dep.notifier).state++;
+        await container.read(provider.future);
 
         expect(container.read(provider), const AsyncData(1));
         expect(container.read(provider.notifier), same(notifier));
@@ -829,14 +745,17 @@ void main() {
     });
 
     test(
-        'Can override AsyncNotifier.updateShouldNotify to change the default filter logic',
-        () {
+        'Can override StreamNotifier.updateShouldNotify to change the default filter logic',
+        () async {
       final provider = factory.simpleTestProvider<Equal<int>>(
-        (ref) => Equal(42),
+        (ref) => Stream.value(Equal(42)),
         updateShouldNotify: (a, b) => a != b,
       );
       final container = ProviderContainer.test();
       final listener = Listener<AsyncValue<Equal<int>>>();
+
+      // Skip the loading
+      await container.listen(provider.future, (previous, next) {}).read();
 
       container.listen(provider, listener.call);
       final notifier = container.read(provider.notifier);
@@ -863,8 +782,11 @@ void main() {
       test('passes in the latest state', () async {
         final container = ProviderContainer.test();
         final provider = factory.simpleTestProvider<int>(
-          (ref) => 0,
+          (ref) => Stream.value(0),
         );
+
+        // Skip the loading
+        await container.listen(provider.future, (previous, next) {}).read();
 
         final sub = container.listen(provider.notifier, (prev, next) {});
 
@@ -926,7 +848,12 @@ void main() {
       test('executes immediately with current state if a state is available',
           () async {
         final container = ProviderContainer.test();
-        final provider = factory.simpleTestProvider<int>((ref) => 1);
+        final provider = factory.simpleTestProvider<int>(
+          (ref) => Stream.value(1),
+        );
+
+        // Skip the loading
+        await container.listen(provider.future, (previous, next) {}).read();
 
         final sub = container.listen(provider.notifier, (prev, next) {});
 
@@ -972,7 +899,7 @@ void main() {
       test('awaits the future resolution if in loading state', () async {
         final container = ProviderContainer.test();
         final provider = factory.simpleTestProvider<int>(
-          (ref) => Future.value(42),
+          (ref) => Stream.value(42),
         );
 
         final sub = container.listen(provider.notifier, (prev, next) {});
@@ -988,49 +915,81 @@ void main() {
     });
   });
 
-  test('supports overrideWith', () {
-    final provider = AsyncNotifierProvider<DeferredAsyncNotifier<int>, int>(
-      () => DeferredAsyncNotifier((ref) => 0),
+  test('supports overrideWith', () async {
+    final provider = StreamNotifierProvider<DeferredStreamNotifier<int>, int>(
+      () => DeferredStreamNotifier((ref) => Stream.value(0)),
     );
     final autoDispose =
-        AsyncNotifierProvider.autoDispose<DeferredAsyncNotifier<int>, int>(
-      () => DeferredAsyncNotifier((ref) => 0),
+        StreamNotifierProvider.autoDispose<DeferredStreamNotifier<int>, int>(
+      () => DeferredStreamNotifier((ref) => Stream.value(0)),
     );
     final container = ProviderContainer.test(
       overrides: [
-        provider.overrideWith(() => DeferredAsyncNotifier((ref) => 42)),
+        provider.overrideWith(
+          () => DeferredStreamNotifier((ref) => Stream.value(42)),
+        ),
         autoDispose.overrideWith(
-          () => DeferredAsyncNotifier((ref) => 84),
+          () => DeferredStreamNotifier((ref) => Stream.value(84)),
         ),
       ],
     );
+
+    // Skip the loading
+    await container.listen(provider.future, (previous, next) {}).read();
+    await container.listen(autoDispose.future, (previous, next) {}).read();
 
     expect(container.read(provider).value, 42);
     expect(container.read(autoDispose).value, 84);
   });
 
-  test('supports family overrideWith', () {
-    final family = AsyncNotifierProvider.family<
-        DeferredFamilyAsyncNotifier<int>, int, int>(
-      () => DeferredFamilyAsyncNotifier((ref) => 0),
+  test('supports family overrideWith', () async {
+    final family = StreamNotifierProvider.family<
+        DeferredFamilyStreamNotifier<int>, int, int>(
+      () => DeferredFamilyStreamNotifier<int>((ref) => Stream.value(0)),
     );
-    final autoDisposeFamily = AsyncNotifierProvider.autoDispose
-        .family<DeferredFamilyAsyncNotifier<int>, int, int>(
-      () => DeferredFamilyAsyncNotifier((ref) => 0),
+    final autoDisposeFamily = StreamNotifierProvider.autoDispose
+        .family<DeferredFamilyStreamNotifier<int>, int, int>(
+      () => DeferredFamilyStreamNotifier<int>((ref) => Stream.value(0)),
     );
     final container = ProviderContainer.test(
       overrides: [
         family.overrideWith(
-          () => DeferredFamilyAsyncNotifier<int>((ref) => 42),
+          () => DeferredFamilyStreamNotifier<int>((ref) => Stream.value(42)),
         ),
         autoDisposeFamily.overrideWith(
-          () => DeferredFamilyAsyncNotifier<int>((ref) => 84),
+          () => DeferredFamilyStreamNotifier<int>(
+            (ref) => Stream.value(84),
+          ),
         ),
       ],
     );
 
+    // Skip the loading
+    await container.listen(family(10).future, (previous, next) {}).read();
+    await container
+        .listen(autoDisposeFamily(10).future, (previous, next) {})
+        .read();
+
     expect(container.read(family(10)).value, 42);
     expect(container.read(autoDisposeFamily(10)).value, 84);
+  });
+
+  group('AutoDispose variant', () {
+    test('can watch autoDispose providers', () async {
+      final dep = Provider.autoDispose((ref) => 0);
+      final provider =
+          StreamNotifierProvider.autoDispose<DeferredStreamNotifier<int>, int>(
+        () => DeferredStreamNotifier((ref) {
+          return Stream.value(ref.watch(dep));
+        }),
+      );
+      final container = ProviderContainer.test();
+
+      // Skip the loading
+      await container.listen(provider.future, (previous, next) {}).read();
+
+      expect(container.read(provider), const AsyncData(0));
+    });
   });
 
   group('modifiers', () {
@@ -1043,13 +1002,11 @@ void main() {
     ) {}
 
     test('provider', () {
-      final provider = AsyncNotifierProvider<DeferredAsyncNotifier<int>, int>(
-        () => DeferredAsyncNotifier((ref) => 0),
+      final provider = StreamNotifierProvider<DeferredStreamNotifier<int>, int>(
+        () => DeferredStreamNotifier((ref) => Stream.value(0)),
       );
 
-      // ignore: avoid_types_on_closure_parameters
       provider.select((AsyncValue<int> value) => 0);
-      // ignore: avoid_types_on_closure_parameters
       provider.selectAsync((int value) => 0);
 
       canBeAssignedToProviderListenable<AsyncValue<int>>(provider);
@@ -1058,19 +1015,17 @@ void main() {
       canBeAssignedToProviderListenable<Future<int>>(provider.future);
       canBeAssignedToRefreshable<Future<int>>(provider.future);
 
-      canBeAssignedToProviderListenable<AsyncNotifier<int>>(provider.notifier);
-      canBeAssignedToRefreshable<AsyncNotifier<int>>(provider.notifier);
+      canBeAssignedToProviderListenable<StreamNotifier<int>>(provider.notifier);
+      canBeAssignedToRefreshable<StreamNotifier<int>>(provider.notifier);
     });
 
     test('autoDispose', () {
       final autoDispose =
-          AsyncNotifierProvider.autoDispose<DeferredAsyncNotifier<int>, int>(
-        () => DeferredAsyncNotifier((ref) => 0),
+          StreamNotifierProvider.autoDispose<DeferredStreamNotifier<int>, int>(
+        () => DeferredStreamNotifier((ref) => Stream.value(0)),
       );
 
-      // ignore: avoid_types_on_closure_parameters
       autoDispose.select((AsyncValue<int> value) => 0);
-      // ignore: avoid_types_on_closure_parameters
       autoDispose.selectAsync((int value) => 0);
 
       canBeAssignedToProviderListenable<AsyncValue<int>>(autoDispose);
@@ -1079,23 +1034,21 @@ void main() {
       canBeAssignedToProviderListenable<Future<int>>(autoDispose.future);
       canBeAssignedToRefreshable<Future<int>>(autoDispose.future);
 
-      canBeAssignedToProviderListenable<AsyncNotifier<int>>(
+      canBeAssignedToProviderListenable<StreamNotifier<int>>(
         autoDispose.notifier,
       );
-      canBeAssignedToRefreshable<AsyncNotifier<int>>(
+      canBeAssignedToRefreshable<StreamNotifier<int>>(
         autoDispose.notifier,
       );
     });
 
     test('family', () {
-      final family = AsyncNotifierProvider.family<
-          DeferredFamilyAsyncNotifier<String>, String, int>(
-        () => DeferredFamilyAsyncNotifier((ref) => '0'),
+      final family = StreamNotifierProvider.family<
+          DeferredFamilyStreamNotifier<String>, String, int>(
+        () => DeferredFamilyStreamNotifier((ref) => Stream.value('0')),
       );
 
-      // ignore: avoid_types_on_closure_parameters
       family(0).select((AsyncValue<String> value) => 0);
-      // ignore: avoid_types_on_closure_parameters
       family(0).selectAsync((String value) => 0);
 
       canBeAssignedToProviderListenable<AsyncValue<String>>(family(0));
@@ -1104,28 +1057,26 @@ void main() {
       canBeAssignedToProviderListenable<Future<String>>(family(0).future);
       canBeAssignedToRefreshable<Future<String>>(family(0).future);
 
-      canBeAssignedToProviderListenable<FamilyAsyncNotifier<String, int>>(
+      canBeAssignedToProviderListenable<FamilyStreamNotifier<String, int>>(
         family(0).notifier,
       );
-      canBeAssignedToRefreshable<FamilyAsyncNotifier<String, int>>(
+      canBeAssignedToRefreshable<FamilyStreamNotifier<String, int>>(
         family(0).notifier,
       );
     });
 
     test('autoDisposeFamily', () {
       expect(
-        AsyncNotifierProvider.autoDispose.family,
-        same(AsyncNotifierProvider.family.autoDispose),
+        StreamNotifierProvider.autoDispose.family,
+        same(StreamNotifierProvider.family.autoDispose),
       );
 
-      final autoDisposeFamily = AsyncNotifierProvider.autoDispose
-          .family<DeferredFamilyAsyncNotifier<String>, String, int>(
-        () => DeferredFamilyAsyncNotifier((ref) => '0'),
+      final autoDisposeFamily = StreamNotifierProvider.autoDispose
+          .family<DeferredFamilyStreamNotifier<String>, String, int>(
+        () => DeferredFamilyStreamNotifier((ref) => Stream.value('0')),
       );
 
-      // ignore: avoid_types_on_closure_parameters
       autoDisposeFamily(0).select((AsyncValue<String> value) => 0);
-      // ignore: avoid_types_on_closure_parameters
       autoDisposeFamily(0).selectAsync((String value) => 0);
 
       canBeAssignedToProviderListenable<AsyncValue<String>>(
@@ -1142,10 +1093,10 @@ void main() {
         autoDisposeFamily(0).future,
       );
 
-      canBeAssignedToProviderListenable<FamilyAsyncNotifier<String, int>>(
+      canBeAssignedToProviderListenable<FamilyStreamNotifier<String, int>>(
         autoDisposeFamily(0).notifier,
       );
-      canBeAssignedToRefreshable<FamilyAsyncNotifier<String, int>>(
+      canBeAssignedToRefreshable<FamilyStreamNotifier<String, int>>(
         autoDisposeFamily(0).notifier,
       );
     });
@@ -1164,25 +1115,22 @@ class Equal<T> {
 
   @override
   int get hashCode => Object.hash(runtimeType, value);
-
-  @override
-  String toString() => 'Equal($value)';
 }
 
-class CtorNotifier extends AsyncNotifier<int> {
+class CtorNotifier extends StreamNotifier<int> {
   CtorNotifier() {
     state;
   }
 
   @override
-  FutureOr<int> build() => 0;
+  Stream<int> build() => Stream.value(0);
 }
 
-class FamilyCtorNotifier extends FamilyAsyncNotifier<int, int> {
+class FamilyCtorNotifier extends FamilyStreamNotifier<int, int> {
   FamilyCtorNotifier() {
     state;
   }
 
   @override
-  FutureOr<int> build(int arg) => 0;
+  Stream<int> build(int arg) => Stream.value(0);
 }
