+++ packages/riverpod/test/new/providers/stream_provider/stream_provider_test.dart
@@ -8,7 +8,7 @@ import 'package:riverpod/riverpod.dart';
 import 'package:riverpod/src/internals.dart' show ProviderElementBase;
 import 'package:test/test.dart';
 
-import '../../utils.dart';
+import '../../../old/utils.dart';
 
 void main() {
   test('supports overrideWith', () {
